web_ui/requirements.txt
# GreytHR Web UI Dashboard - Dependencies
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0

# Template engine
jinja2==3.1.2

# File operations
aiofiles==23.2.1
python-multipart==0.0.6

# Configuration management
pyyaml==6.0.1

# WebSocket support
websockets==12.0

# Async HTTP client (for potential external API calls)
httpx==0.25.2

# Date/time handling
python-dateutil==2.8.2

# Development dependencies (optional)
# Uncomment if needed for development
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0

# Security (if authentication is added later)
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# Monitoring/observability (optional)
# prometheus-client==0.19.0
